restore the header file & separate logo
